Set Region To Fake Global
When executing the ansible s3 module, it seems that bucket will be
"registered" to fake region, even though AWS considers buckets to be of
region type global. If we execute an operation against a bucket
registered as us-east-1, but with AWS_REGION set to us-east-2, we get
the error detailed in the below issue. Setting region attribute of s3
module to "global" mitigates this issue.

https://github.com/ansible/ansible/issues/29397